entregable3 primera parte

+++ src/main/java/com/itsqmet/nucleo/servicio/IMatriculaServicio.java
@@ -1,5 +1,6 @@
 package com.itsqmet.nucleo.servicio;
 
+
 import com.itsqmet.nucleo.modelo.Matricula;
 
 import java.util.List;
@@ -12,11 +13,11 @@ public interface IMatriculaServicio {
     public List<Matricula> listarMatriculas();
 
     // Método para obtener una matrícula por su ID
-    public Matricula obtenerMatriculaPorId(Long id);
+    public Matricula obtenerMatriculaPorId(Integer id);
 
     // Método para actualizar los datos de una matrícula
     public void actualizarMatricula(Matricula matriculaActualizada);
 
     // Método para eliminar una matrícula por su ID
-    public void eliminarMatricula(Long id);
+    public void eliminarMatricula(Integer id);
 }
